backend/requirements.txt
django
psycopg2-binary
bcrypt
cffi
cryptography
ecdsa
passlib
pyasn1
pycparser
python-jose[cryptography]
python-multipart
rsa
six
dotenv
dj-database-url
python-dotenv
stripe
supabase
supabase-auth
django-environ
